requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
xmltodict==0.13.0
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
rapidfuzz==3.6.1
sentence-transformers==2.2.2
supabase==2.0.0
celery==5.3.4
redis==5.0.1
fastapi-limiter==0.1.6
slowapi==0.1.9
cryptography==41.0.7
stripe==7.0.0
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
httpx==0.24.1

